wip for days 14,15 and 16
